refactor(internal/librarian): move interfaces and isURL to command.go
Move the GitHubClient and ContainerClient interfaces and isURL function
from librarian.go to command.go to better organize code structure.

This setups a followup PR that moves command initialization to
librarian.go, inside of relying on init functions.
